+++ src/Syringe.Core/Runner/TestFileRunner.cs
@@ -5,10 +5,10 @@ using System.Threading.Tasks;
 using Syringe.Core.Http;
 using Syringe.Core.Http.Logging;
 using Syringe.Core.Logging;
-using Syringe.Core.Repositories;
 using Syringe.Core.Runner.Assertions;
 using Syringe.Core.Tests;
 using Syringe.Core.Tests.Results;
+using Syringe.Core.Tests.Results.Repositories;
 using HttpResponse = Syringe.Core.Http.HttpResponse;
 
 namespace Syringe.Core.Runner
