+++ src/Syringe.Core/Tests/Repositories/Xml/Writer/TestFileWriter.cs
@@ -1,9 +1,8 @@
 using System.Text;
 using System.Xml;
 using System.Xml.Linq;
-using Syringe.Core.Tests;
 
-namespace Syringe.Core.Repositories.XML.Writer
+namespace Syringe.Core.Tests.Repositories.Xml.Writer
 {
     public class TestFileWriter : ITestFileWriter
     {
