+++ src/Syringe.Service/DependencyResolution/DefaultRegistry.cs
@@ -25,11 +25,11 @@ using StructureMap.Graph;
 using Syringe.Core.Configuration;
 using Syringe.Core.Environment;
 using Syringe.Core.IO;
-using Syringe.Core.Repositories;
-using Syringe.Core.Repositories.MongoDB;
-using Syringe.Core.Repositories.XML;
-using Syringe.Core.Repositories.XML.Reader;
-using Syringe.Core.Repositories.XML.Writer;
+using Syringe.Core.Tests.Repositories;
+using Syringe.Core.Tests.Repositories.Xml;
+using Syringe.Core.Tests.Repositories.Xml.Reader;
+using Syringe.Core.Tests.Repositories.Xml.Writer;
+using Syringe.Core.Tests.Results.Repositories;
 using Syringe.Service.Api.Hubs;
 using Syringe.Service.Parallel;
 using WebApiContrib.IoC.StructureMap;
