+++ src/Syringe.Service/Parallel/ParallelTestFileQueue.cs
@@ -8,12 +8,12 @@ using System.Threading.Tasks;
 using RestSharp;
 using Syringe.Core.Configuration;
 using Syringe.Core.Http;
-using Syringe.Core.Repositories;
-using Syringe.Core.Repositories.XML.Reader;
 using Syringe.Core.Runner;
 using Syringe.Core.Tasks;
 using Syringe.Core.Tests;
+using Syringe.Core.Tests.Repositories.Xml.Reader;
 using Syringe.Core.Tests.Results;
+using Syringe.Core.Tests.Results.Repositories;
 
 namespace Syringe.Service.Parallel
 {
