+++ src/Syringe.Tests/Integration/Xml/TestFileRunnerTests.cs
@@ -5,12 +5,11 @@ using Moq;
 using NUnit.Framework;
 using RestSharp;
 using Syringe.Core.Http;
-using Syringe.Core.Repositories;
-using Syringe.Core.Repositories.MongoDB;
-using Syringe.Core.Repositories.XML.Reader;
 using Syringe.Core.Runner;
 using Syringe.Core.Tests;
+using Syringe.Core.Tests.Repositories.Xml.Reader;
 using Syringe.Core.Tests.Results;
+using Syringe.Core.Tests.Results.Repositories;
 using Syringe.Tests.StubsMocks;
 
 namespace Syringe.Tests.Integration.Xml
