+++ src/Syringe.Tests/StubsMocks/TestCaseSessionRepositoryMock.cs
@@ -1,8 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using Syringe.Core.Repositories;
 using Syringe.Core.Tests.Results;
+using Syringe.Core.Tests.Results.Repositories;
 
 namespace Syringe.Tests.StubsMocks
 {
