+++ src/Syringe.Tests/Unit/Core/Tests/Repositories/Xml/Reader/XmlExamples/full.xml
