+++ src/Syringe.Tests/Unit/Core/Tests/Repositories/Xml/Reader/XmlExamples/single-test.xml
