+++ src/Syringe.Tests/Unit/Core/Tests/Repositories/Xml/Writer/XmlExamples/headers.xml
